==> notebooks/01_exploracion_youtube_api.ipynb <==
# 01 Exploracion YouTube Api

Este notebook se encarga de explorar los datos de YouTube Data API v3
y revisar qué datos tenemos disponible y cómo están estruturados

import sys
print(sys.executable)


#Sí sale True indica que el archivo .env fue encontrado y cargado con todas las variables que tiene dentro
import os
from dotenv import load_dotenv

load_dotenv()

#Recurso - channels - vista rápida
import requests
import os

API_KEY = os.getenv("YOUTUBE_API_KEY")
CHANNEL_ID = os.getenv("YOUTUBE_CHANNEL_ID")

url = "https://www.googleapis.com/youtube/v3/channels"

params = {
    "part": "snippet,statistics,contentDetails",
    "id": CHANNEL_ID,
    "key": API_KEY
}

response = requests.get(url, params=params)
response.raise_for_status()

channel_data = response.json()
channel_data

# Guardar el JSON crudo de channels para analizar los datos
# Solo se guarda una parte para revisión, en la parte de extracción se usará la misma ruta para los datos reales de produccion

import os
import json

# Ruta base donde guardaremos el raw
base_path = "/Users/angelgarciachanga/repositorios/publico/youtube-v3-data-pipeline/data/raw/youtube"

# Asegurar que la carpeta existe
#os.makedirs(base_path, exist_ok=True)

# Ruta final del archivo
output_path = f"{base_path}/channels.json"

# Guardar el JSON crudo de channels
with open(output_path, "w", encoding="utf-8") as f:
    json.dump(channel_data, f, indent=2, ensure_ascii=False)

output_path


#extraer y guardar el ID que es la llave de todos los videos del canal.
uploads_playlist_id = channel_data["items"][0]["contentDetails"]["relatedPlaylists"]["uploads"]
uploads_playlist_id

#Recurso - playlistItems - vista rápida
import requests
import os

API_KEY = os.getenv("YOUTUBE_API_KEY")

url = "https://www.googleapis.com/youtube/v3/playlistItems"

params = {
    "part": "snippet,contentDetails",
    "playlistId": uploads_playlist_id,
    "maxResults": 5,   # solo para probar
    "key": API_KEY
}

response = requests.get(url, params=params)
response.raise_for_status()

playlist_items = response.json()
playlist_items




# Guardar el JSON crudo de playlist_items para analizar los datos
# Solo se guarda una parte para revisión, en la parte de extracción se usará la misma ruta para los datos reales de produccion

import os
import json

# Ruta base del raw de YouTube
base_path = "/Users/angelgarciachanga/repositorios/publico/youtube-v3-data-pipeline/data/raw/youtube"

# Crear carpetas si no existen
os.makedirs(base_path, exist_ok=True)

# Ruta final del archivo
output_path = f"{base_path}/playlist_items.json"

# Guardar JSON crudo
with open(output_path, "w", encoding="utf-8") as f:
    json.dump(playlist_items, f, indent=2, ensure_ascii=False)

output_path


#Recurso - playlist - vista rápida
import os
import requests

API_KEY = os.getenv("YOUTUBE_API_KEY")
CHANNEL_ID = os.getenv("YOUTUBE_CHANNEL_ID")

url = "https://www.googleapis.com/youtube/v3/playlists"

params = {
    "part": "snippet,contentDetails",
    "channelId": CHANNEL_ID,
    "maxResults": 5,  # vista rápida
    "key": API_KEY
}

response = requests.get(url, params=params)
response.raise_for_status()

playlists = response.json()
playlists


# Guardar el JSON crudo de playlists para analizar los datos
# Solo se guarda una parte para revisión, en la parte de extracción se usará la misma ruta para los datos reales de produccion

import os
import json

# Ruta base del raw de YouTube
base_path = "/Users/angelgarciachanga/repositorios/publico/youtube-v3-data-pipeline/data/raw/youtube"

# Crear carpetas si no existen
os.makedirs(base_path, exist_ok=True)

# Ruta final del archivo
output_path = f"{base_path}/playlists.json"

# Guardar JSON crudo
with open(output_path, "w", encoding="utf-8") as f:
    json.dump(playlists, f, indent=2, ensure_ascii=False)

output_path

# Extraer videoId desde playlist_items
video_ids = [
    item["contentDetails"]["videoId"]
    for item in playlist_items["items"]
]

video_ids

#Recurso - videos - vista rápida
import requests
import os

API_KEY = os.getenv("YOUTUBE_API_KEY")

url = "https://www.googleapis.com/youtube/v3/videos"

params = {
    "part": "snippet,statistics,contentDetails",
    "id": ",".join(video_ids),
    "key": API_KEY
}

response = requests.get(url, params=params)
response.raise_for_status()

videos_data = response.json()
videos_data


# Guardar el JSON crudo de videos para analizar los datos

import os
import json

# Ruta base del raw de YouTube
base_path = "/Users/angelgarciachanga/repositorios/publico/youtube-v3-data-pipeline/data/raw/youtube"

# Crear carpetas si no existen
os.makedirs(base_path, exist_ok=True)

# Ruta final del archivo
output_path = f"{base_path}/videos.json"

# Guardar JSON crudo
with open(output_path, "w", encoding="utf-8") as f:
    json.dump(videos_data, f, indent=2, ensure_ascii=False)

output_path
