==> notebooks/05_transform_playlist_items_uploads(x).ipynb <==
# 05 – Transformación: playlist_items_uploads

Esta tabla contine todas los videos del canal con una playlist igual para todas (uploads) la cual es creada por el mismo canal.
No contiene métricas ni snapshots, solo la estructura editorial
del canal.


from pathlib import Path
import json
import pandas as pd
from datetime import datetime

PROJECT_ROOT = Path.cwd().parent
RAW_PATH = PROJECT_ROOT / "data" / "raw" / "youtube"
PROCESSED_PATH = PROJECT_ROOT / "data" / "processed" / "youtube"

PROCESSED_PATH.mkdir(parents=True, exist_ok=True)


with open(RAW_PATH / "playlist_items_uploads.json", "r", encoding="utf-8") as f:
    playlist_items_raw = json.load(f)

len(playlist_items_raw["items"])


from datetime import datetime, timezone

# Timestamp explícitamente en UTC
EXTRACTED_AT = datetime.now(timezone.utc)

EXTRACTED_AT

rows = []

for item in playlist_items_raw["items"]:
    rows.append({
        "playlist_id": item["snippet"]["playlistId"],
        "video_id": item["contentDetails"]["videoId"],
        "position": item["snippet"]["position"],
        "added_at": item["snippet"]["publishedAt"],
        "extracted_at": EXTRACTED_AT
    })

df_playlist_items_uploads = pd.DataFrame(rows)
df_playlist_items_uploads.head()


df_playlist_items.dtypes

# El added_at viene de la API como string ISO (ya en UTC normalmente con Z), debes convertirlo:
df_playlist_items_uploads["added_at"] = (
    pd.to_datetime(df_playlist_items_uploads["added_at"], utc=True)
)

df_playlist_items.shape

> ⚠️ Nota -> NO SE GUARDARÁ df_playlist_items_uploads COMO PARQUET PORQUE NO SE UTILIZARÁ DEBIDO A QUE LA PLAYLIST QUE TIENE ES LA MISMA PARA TODAS LOS VIDEOS, EN VEZ DE ESO SE UTILIZARÁ df_playlist_items_manual
